requirements.txt
numpy
pandas
torch
torchvision
torchaudio
pytorch-lightning
nemo-toolkit[all]
hydra-core
lightning
omegaconf
lhotse
einops
transformers
sentencepiece
jiwer
librosa
webdataset
pyannote.audio
editdistance